INIFile and INISection: Inherit from OrderedDictionary
By using an OrderedDictionary, the writer will write the sections and keys
in the same order they were read.

But as the Dictionary-class doesn't allow to override its methods, an
own implementation of IDictionary was done to create the OrderedDictionary.
Therefore the caller doesn't get a Dictionary anymore, but only a IDictionary object.