Remove the Markdown notice for NuGet
NuGet doesn't support tables in their markdown, therefore a notice
and a link to the Github was added.
But as the README growed, I noticed that NuGet only supports 8000 bytes (!)
for the README/Documentation.
So I had to create a separate version of the README for NuGet anyway.
Therefore we don't need the notice in this repo anymore.